moved bridge declarations to dedicated folder
