+++ relays/bin-substrate/src/bridges/rialto_millau/mod.rs
@@ -14,16 +14,9 @@
 // You should have received a copy of the GNU General Public License
 // along with Parity Bridges Common.  If not, see <http://www.gnu.org/licenses/>.
 
-//! Substrate-to-substrate relay entrypoint.
+//! Declaration of all bridges between Rialto and Millau.
 
-#![warn(missing_docs)]
-
-mod bridges;
-mod chains;
-mod cli;
-
-fn main() {
-	let command = cli::parse_args();
-	let run = command.run();
-	async_std::task::block_on(run);
-}
+pub mod millau_headers_to_rialto;
+pub mod millau_messages_to_rialto;
+pub mod rialto_headers_to_millau;
+pub mod rialto_messages_to_millau;
