+++ relays/bin-substrate/src/bridges/rococo_wococo/mod.rs
@@ -14,16 +14,11 @@
 // You should have received a copy of the GNU General Public License
 // along with Parity Bridges Common.  If not, see <http://www.gnu.org/licenses/>.
 
-//! Substrate-to-substrate relay entrypoint.
+//! Declaration of all bridges between Rococo Bridge Hub and Wococo Bridge Hub.
 
-#![warn(missing_docs)]
-
-mod bridges;
-mod chains;
-mod cli;
-
-fn main() {
-	let command = cli::parse_args();
-	let run = command.run();
-	async_std::task::block_on(run);
-}
+pub mod bridge_hub_rococo_messages_to_bridge_hub_wococo;
+pub mod bridge_hub_wococo_messages_to_bridge_hub_rococo;
+pub mod rococo_headers_to_bridge_hub_wococo;
+pub mod rococo_parachains_to_bridge_hub_wococo;
+pub mod wococo_headers_to_bridge_hub_rococo;
+pub mod wococo_parachains_to_bridge_hub_rococo;
