+++ relays/bin-substrate/src/chains/mod.rs
@@ -16,23 +16,6 @@
 
 //! Chain-specific relayer configuration.
 
-pub mod bridge_hub_rococo_messages_to_bridge_hub_wococo;
-pub mod bridge_hub_wococo_messages_to_bridge_hub_rococo;
-pub mod millau_headers_to_rialto;
-pub mod millau_headers_to_rialto_parachain;
-pub mod millau_messages_to_rialto;
-pub mod millau_messages_to_rialto_parachain;
-pub mod rialto_headers_to_millau;
-pub mod rialto_messages_to_millau;
-pub mod rialto_parachain_messages_to_millau;
-pub mod rialto_parachains_to_millau;
-pub mod rococo_headers_to_bridge_hub_wococo;
-pub mod rococo_parachains_to_bridge_hub_wococo;
-pub mod westend_headers_to_millau;
-pub mod westend_parachains_to_millau;
-pub mod wococo_headers_to_bridge_hub_rococo;
-pub mod wococo_parachains_to_bridge_hub_rococo;
-
 mod millau;
 mod rialto;
 mod rialto_parachain;
