+++ relays/bin-substrate/src/cli/relay_headers.rs
@@ -20,13 +20,17 @@ use sp_core::Pair;
 use structopt::StructOpt;
 use strum::{EnumString, EnumVariantNames, VariantNames};
 
-use crate::chains::{
-	millau_headers_to_rialto::MillauToRialtoCliBridge,
-	millau_headers_to_rialto_parachain::MillauToRialtoParachainCliBridge,
-	rialto_headers_to_millau::RialtoToMillauCliBridge,
-	rococo_headers_to_bridge_hub_wococo::RococoToBridgeHubWococoCliBridge,
-	westend_headers_to_millau::WestendToMillauCliBridge,
-	wococo_headers_to_bridge_hub_rococo::WococoToBridgeHubRococoCliBridge,
+use crate::bridges::{
+	rialto_millau::{
+		millau_headers_to_rialto::MillauToRialtoCliBridge,
+		rialto_headers_to_millau::RialtoToMillauCliBridge,
+	},
+	rialto_parachain_millau::millau_headers_to_rialto_parachain::MillauToRialtoParachainCliBridge,
+	rococo_wococo::{
+		rococo_headers_to_bridge_hub_wococo::RococoToBridgeHubWococoCliBridge,
+		wococo_headers_to_bridge_hub_rococo::WococoToBridgeHubRococoCliBridge,
+	},
+	westend_millau::westend_headers_to_millau::WestendToMillauCliBridge,
 };
 use relay_utils::metrics::{GlobalMetrics, StandaloneMetric};
 use substrate_relay_helper::finality::SubstrateFinalitySyncPipeline;
